requirements.txt
click>=8.0.0
paramiko>=3.0.0
PyYAML>=6.0
pyusb>=1.2.0
python-dotenv>=1.0.0
